+++ .env.dist
@@ -1,0 +1,7 @@
+BOT_TOKEN=576281325955420:AAHJJlmnd488vdDZBod_vgf5849aiUX-F9Jkf558QDnc
+SEND_REPORT_TO_OWNER=False
+POSTGRES_HOST=db
+POSTGRES_USER="postgres"
+POSTGRES_DB="cnc_bot"
+POSTGRES_PORT="5432"
+POSTGRES_PASSWORD="postgres"
