--- PK.txt
@@ -1,36 +1,0 @@
-#8 = 0 (X)
-#9 = 0 (Y)
-#10 = #1 - #6
-#11 = #2 - #6
-#12 = #7 * #6
-#13 = #11 / #10
-#14 = #12 * #13
-N2
-G90 G0 X#8 Y#9
-#8 = #8 + #12
-#9 = #9 + #14
-Z[#4 + #5]
-G01 X-#8 Z#4
-#4 = #4 - #5
-X#8 Z#4
-N1
-G90 X#8
-Y#9
-X-#8
-Y-#9
-X#8
-G91 Y#9
-IF [#8 EQ #10] GOTO 3
-#8 = #8 + #12
-#9 = #9 + #14
-IF [#8 LT #10] GOTO 1
-#8 = #10
-#9 = #11
-IF [#8 EQ #10] GOTO 1
-N3
-#8 = 0
-#9 = 0
-#4 = #4 - #5
-IF [#4 GT #3] GOTO 2
-#4 = 0
-G0 Z200
