+++ README.md
@@ -1,12 +1,15 @@
-- 👋 Hi, I’m @Artyom-Murzinov
-- 👀 I’m interested in Металлообработка, автоматизация!
-- 🌱 I’m currently learning Python
-- 💞️ I’m looking to collaborate on проектам в области металлоообработки и ее автоматизации и повышению ее производительности
-- 📫 How to reach me:
-- 😄 Pronouns: ...
-- ⚡ Fun fact: ...
+# Бот для помощи в составлении управляющих программ ЧПУ станков
+ -
 
-<!---
-Artyom-Murzinov/Artyom-Murzinov is a ✨ special ✨ repository because its `README.md` (this file) appears on your GitHub profile.
-You can click the Preview link to take a look at your changes.
---->
+# Быстрый старт
+ -
+
+# Технологии
+ - telebot # https://github.com/eternnoir/pyTelegramBotAPI.git
+
+# Дорожная карта 
+ - структура проекта 
+ - использование скрытого окружения для деплоя
+ - асинхроность?
+
+
