--- RK.txt
@@ -1,65 +1,0 @@
-#10 = SIN[#5]* #6
-#11 = COS[#5]* #6
-#12 = #5 + #2
-#13 = SIN[#12]* #6
-#14 = COS[#12]* #6
-#15 = #6
-#16 = #7* #9
-#17 = #16
-#1 = #1 - #7
-#20 = #16* 2
-#21 = ASIN[#20 / #6]
-#22 = #21 + #5
-#23 = SIN[#22]* #6
-#24 = COS[#22]* #6
-#8 = #8 / 2
-G0 X#11 Y#10
-Z#4
-N1
-#15 = #6 + #17
-#10 = SIN[#5]* #15
-#11 = COS[#5]* #15
-G1 X#11 Y#10
-#13 = SIN[#12]* #15
-#14 = COS[#12]* #15
-G3 X#14 Y#13 I-#11 J-#10
-#18 = SIN[#12]* #17
-#19 = COS[#12]* #17
-#15 = #6 - #17
-#13 = SIN[#12]* #15
-#14 = COS[#12]* #15
-G3 X#14 Y#13 I-#19 J-#18
-#10 = SIN[#5]* #15
-#11 = COS[#5]* #15
-G2 X#11 Y#10 I-#14 J-#13
-#15 = #6+ #17
-#10 = SIN[#5]* #15
-#11 = COS[#5]* #15
-#18 = SIN[#5]* #17
-#19 = COS[#5]* #17
-G3 X#11 Y#10 I+#19 J+#18
-IF [#17 EQ #1] GOTO 2
-#17 = #17 + #16
-IF [#17 LT #1] GOTO1
-#17 = #1
-IF [#17 EQ #1] GOTO 1
-N2
-#15 = #6
-#17 = #16
-IF [#4 EQ #3] GOTO 3
-N10
-#10 = SIN[#5]* #6
-#11 = COS[#5]* #6
-G0 X#11 Y#10
-IF [#4 EQ #3] GOTO 12
-Z#4
-#4 = #4 - #8
-IF [#4 LT #3] THEN #4 = #3
-G3 X#24 Y#23 Z#4 I-#11 J-#10
-#4 = #4 - #8
-IF [#4 LT #3] THEN #4 = #3
-G2 X#11 Y#10 Z#4 I-#24 J-#23
-IF [#4 GE #3] GOTO 1
-N12
-G0 Z200
-M30
