+++ src/main.py
@@ -1,0 +1,4 @@
+from core.bot import bot
+import asyncio
+
+asyncio.run(bot.polling())
