+++ src/static/KK.txt
@@ -1,0 +1,35 @@
+#7 = #5 * #6
+#8 = #1/2 - #5
+#9 = #4 - #3
+#10 = #7
+#11 = #4
+G92 X0 Y0
+F600
+N1
+G0 X0 Y0
+Z#11
+G1X#10
+G3I-#10Z#9
+#10=#10+#7
+IF[#10GE#8]GOTO3
+N2
+G1X#10
+G3I-#10
+IF[#10EQ#8]GOTO4
+#10=#10+#7
+IF[#10LT#8]GOTO2
+N3
+#10 = #8
+IF[#10EQ#8]GOTO2
+N4
+#10 = #7
+#11 = #9
+G0Z#4
+IF[#9EQ#2]GOTO5
+#9 = #9-#3
+IF[#9GT#2]GOTO1
+#9 = #2
+IF[#9EQ#2]GOTO1
+N5
+#9 = #4
+G0Z50
