+++ src/static/dictionary.py
@@ -1,0 +1,19 @@
+word = ["Введите диаметр отверстия: ", "Введите глубину отверстия, пишется с отрицательным значением: ",
+        "Введите радиус фрезы: ", "Введите шаг фрезерования: ",
+        "Поверхность детали: ", "Коэффициент перекрытия фрезы, 1-полный диаметр: "]
+word_2 = ["Введите длину кармана: ", "Введите ширину кармана: ", "Введите глубину кармана: ",
+          "Поверхность детали: ", "Введите шаг фрезерования: ", "Введите радиус фрезы: ",
+          "Коэффициент перекрытия фрезы: "]
+word_3 = ["Введите ширину кармана: ", "Введите длину кармана в градусах: ", "Введите глубину кармана: ",
+          "Поверхность детали: ", "Начальный угол: ", "Средний радиус паза: ", "Введите радиус фрезы: ",
+          "Введите шаг фрезерования: ", "Коэффициент перекрытия фрезы: "]
+word_4 = ['Введите скорость Vc', 'Введите диаметр фрезы D', 'Введите число зубев Z', 'Введите съем на зуб Fz']
+user = {}
+
+
+async def greeting_text(user_full_name) -> str:
+    return (
+        f'Привет, {user_full_name}!\n\n'
+        f'Я бот помогающий составить управяющю программу для станка ЧПУ\n\n'
+        f'Ознакомся с TODO job description\n\n'
+    )
